map: _getSafeKey(): Generate ID after classes
Since HTML5 (REC), the “id” attribute of elements may contain the
dot (“.”) character which is also used to separate class names in
the safe key.  This modification allows to tell apart a DOM object
representing an element with ID “foo”, and classes “bar” and
“baz” (formerly: "…#foo.bar.baz"; now: "….bar.baz#foo"), from
a DOM object representing an element with ID “foo.bar” and class
“baz” (formerly: "…#foo.bar.baz"; now: "….baz#foo.bar").

Although there are no restrictions on the HTML “class” attribute
value, it is considered less likely that the value would contain
the “#” character as the CSS Selectors Level 3 (REC) Specification
allows it only escaped in class selectors.