ResourceBuilder: Fixed JSdoc-stripping of complex code
\preg_replace_callback() in stripJSdoc() can easily run into 
pcre.backtrack_limit with more complex code (like object.js).  No error 
is issued in that case, but \preg_replace_callback() returns NULL. This 
can happen only when NOT in debug mode as we do not strip JSdoc in debug 
mode.

As a workaround, in such a case we advise in an output comment to 
increase the setting, and do not attempt to strip JSdoc.

See also:

https://www.php.net/manual/en/function.preg-replace-callback.php#75963 
(by Sjon@hortensius.net)

https://www.php.net/manual/en/function.preg-replace-callback.php#98721 
(by chris@ocproducts.com)